Avoid a C5219 warning from MSVC 2019
The warning was: "performance_counter_policy.cpp(59): warning C5219:
implicit conversion from 'const __int64' to 'long double', possible loss
of data".

+++ breeze/time/windows_only/brz/performance_counter_policy.cpp
@@ -55,7 +55,7 @@ performance_counter_policy::duration_type
 performance_counter_policy::resolution() const
 {
     return duration_type(
-         static_cast< long double >( duration_type::period::den ) /
+        duration_type::period::den /
             m_frequency /
             duration_type::period::num
         ) ;
